requirements.txt
pytesseract
pillow
PyPDF2
python-docx
